backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
motor
openai
